+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Bug Fixes
 ---------
 * Fix incorrect metadata for compact counter tables (PYTHON-1100)
 * Call ConnectionException with correct kwargs (PYTHON-1117)
+* Fix an ExecutionProfile can be added to multiple cluster objects (PYTHON-1111)
 3.18.0
 ======
 May 27, 2019